Move examples/model.py to utilities_examples.py
