+++ examples_artificial_data/energy_ratio/energy_ratio_vs_wd_for_single_df.py
@@ -21,7 +21,7 @@ from flasc.energy_ratio.energy_ratio_input import EnergyRatioInput
 from flasc.dataframe_operations import \
     dataframe_manipulations as dfm
 from flasc import floris_tools as fsatools
-from flasc.examples.models import load_floris_artificial as load_floris
+from flasc.utilities_examples import load_floris_artificial as load_floris
 
 
 def load_data():
