+++ examples_artificial_data/wake_steering_design/01_analyze_pP_values.py
@@ -18,7 +18,7 @@ import seaborn as sns
 
 from flasc.wake_steering.lookup_table_tools import get_yaw_angles_interpolant
 
-from flasc.examples.models import load_floris_artificial as load_floris
+from flasc.utilities_examples import load_floris_artificial as load_floris
 from _local_helper_functions import optimize_yaw_angles, evaluate_optimal_yaw_angles
