+++ flasc/__init__.py
@@ -9,7 +9,6 @@ __version__ = '1.0'
 from pathlib import Path
 
 from . import (
-    examples,
     dataframe_operations,
     energy_ratio,
     model_estimation,
@@ -21,6 +20,7 @@ from . import (
     optimization,
     time_operations,
     utilities,
+    utilities_examples,
     visualization,
     timing_tests
 )
