--- flasc/examples/__init__.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Model estimation module for FLORIS SCADA Analysis repository."""
-
-__author__ = """Bart Doekemeijer"""
-__email__ = 'bart.doekemeijer@nrel.gov'
-__version__ = '0.1.0'
-
-from pathlib import Path
-
-from . import models
